% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_matches.R
\name{closest_matches}
\alias{closest_matches}
\title{Find the closest n variables to your search query}
\usage{
closest_matches(a, category = NULL, n = 10, variable.names = FALSE, ...)
}
\arguments{
\item{a}{Your query, a character vector}

\item{category}{If you want to search within a particular category (hdng_names_cats)}

\item{n}{The amount of variables you want to return.}

\item{variable.names}{= FALSE. Whether you want to include the description or the abbreviation.}

\item{...}{Arguments to be passed on to \code{stringdist::stringdistmatrix}.}
}
\value{
a matrix of n rows and length(a) columns
}
\description{
This function returns a matrix of n rows and length(a) columns with all names
of the closest matches of the variables to the user's query,
allowing the user to efficiently gather variables for a data query.
}
\examples{
closest_matches(c("Verkeer", "Bevolking"), n = 20, variable.names = T)
closest_matches(c("Aaantal vrachtwagens", "Aantal autos", "Schoolgaande kinderen"), n = 30)
}
\seealso{
hdng::hdng_names_cats
}
