% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdng_names_cats.R
\name{hdng_names_cats}
\alias{hdng_names_cats}
\title{Explore categories and variable names}
\usage{
hdng_names_cats(cat, from = 1000, to = 2000, show.only.available = TRUE)
}
\arguments{
\item{cat}{A category, defaults to NULL, gives a list of all categories}

\item{from}{Time frame for your query (start)}

\item{to}{Time frame for your query (end)}

\item{show.only.available}{Defaults to TRUE. Filters only variables which are available at least once.}
}
\value{
a data.frame with var name, category, var code, and years of availability
}
\description{
A function that helps the user explore the available variables.
The function returns all categories on an empty call, and
returns all variables in a category on a call with a category (typo robust)
allowing the user to efficiently gather variables for a data query.
}
\examples{
hdng_names_cats()
hdng_names_cats("Beroepen", from = 1850, to = 1900)
hdng_names_cats(c("Politiek", "Woningen"))
}
\seealso{
hdng::hdng_names_cats
}
